feat!: use kv reader instead of generic logical reader (#13)
This is a breaking change for users that are using KVv1 secret storage and were relying on the fact that the internal reader indiscriminately prepend kv/data/ to every secret spec (which was a bug, only exhibiting "correct" behavior if the secret backend in question was KVv2). This change uses the native vault KVv1 and KVv2 readers to properly read the path supplied in the secret spec. Functionality should be congruent to how one uses vault kv get, i.e kv/data will be implicitly prepended to the path if and only if the backend is also truly KVv2.

* feat!: use kv reader instead of generic logical reader

this is a breaking change for users that are using kv-v1 secret storage

* chore: bump supported go versions

* fix: allow nonfree software in nix shell

* fix: make vault-docker script posix-compliant

* docs: update README

* test: refactor patch tests to use string reader

* test: run tests in parallel